# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist:
    post:
      operationId: ArtistsController_create
      summary: Add new artist
      description: Add new artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. body does not contain required fields
      tags:
        - Artists
    get:
      operationId: ArtistsController_findAll
      summary: Gets all artists
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
      tags:
        - Artists
  /artist/{id}:
    get:
      operationId: ArtistsController_findOne
      summary: Get single artist by id
      description: Get single artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Artists
    put:
      operationId: ArtistsController_update
      summary: Update artist information
      description: Update artist information by UUID
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: '"Artist"'
                example:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  name: Test Artist
                  grammy: true
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                    example: Test artist
                  grammy:
                    type: boolean
                    example: 'false'
        '400':
          description: Bad request. ArtistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Artists
    delete:
      operationId: ArtistsController_remove
      summary: Delete artist
      description: Deletes artist by ID.
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            type: string
      responses:
        '204':
          description: The artist has been deleted
        '400':
          description: Bad request. ArtistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags:
        - Artists
...
  /favs/{entity}/{id}:
    post:
      operationId: FavoritesController_Add
      summary: add album/artist/track to favorite
      parameters:
        - name: entity
          required: true
          in: path
          schema:
            type: string
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Successful operation
        '400':
          description: Bad request. Id is invalid (not uuid)
        '422':
          description: Not found
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_remove
      summary: delete album/artist/track from favorite
      parameters:
        - name: entity
          required: true
          in: path
          schema:
            type: string
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Successful operation
        '400':
          description: Bad request. Id is invalid (not uuid)
        '422':
          description: Not found
      tags:
        - Favorites
info:
  title: Home Library Service
  description: Home music library service
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
...
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: name
          nullable: false
        grammy:
          type: boolean
          description: grammy
          nullable: false
      required:
        - name
        - grammy
    Artist:
      type: object
      properties:
        id:
          type: string
          description: Artist identifier
          format: uuid
        name:
          type: string
          description: Artist name
          example: Test user
        grammy:
          type: boolean
          description: Grammy
          example: 'false'
      required:
        - id
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: name
          nullable: false
        grammy:
          type: boolean
          description: grammy
          nullable: false
...